Reimplemented glob to regex translation
- improved handling of escaped characters (e.g. a\[.abc)
- improved finding matching brackets and braces
- creates regular expressions to match numerical ranges
- unit-tests to test glob to regex translation

The regex for numerical ranges can be switched between three modes:

- AS_IS: Equivalent to the previous implementation
- ZEROS: Allow any number of leading zeros
- JUSTIFIED: Implement ranges similar to bash

The final mode depends on the outcome of
https://github.com/editorconfig/editorconfig/issues/371

Currently active: JUSTIFIED (the decision I would prefer :-))

On braces and brackets:

'{alpha,[a,]beta}' should become '%(alpha\|\[a\|\]beta\)' as described
in the bash man page.
So if we scan brackets, that are inside braces, the comma should be
handled as a separator for the braces. To prevent this escape the
comma with a backslash:
'{alpha,[a\,]beta}' becomes '%(alpha\|[a,]beta\)'.