Fix fnmatch adjacent nested curly braces (#33)
* Fix fnmatch adjacent nested curly braces

If an fnmatch expression had adjacent nested curly braces, the braces
would be treated as literals.  For example:

    translate('{a,{b,c}}') == '(?s)\\{a\\,\\{b\\,c\\}\\}\\Z'
    translate('{{a,b},c}') == '(?s)\\{\\{a\\,b\\}\\,c\\}\\Z'

They should be translated as:

    translate('{a,{b,c}}') == '(?s)(?:a|(?:b|c))\\Z'
    translate('{{a,b},c}') == '(?s)(?:(?:a|b)|c)\\Z'

This can occur when matching a mix of file names and extensions.  For
example, I encountered the issue for `{CMakeLists.txt,*.{c,cmake,sh}}`.

The issue occurs because LEFT_BRACE and RIGHT_BRACE include the
preceding character in the match, which causes the match to skip the
second adjacent curly brace, which causes the match count to differ.  To
avoid this, use a zero-width assertion (negative look-behind in this
case).

Note: This makes adjacent nested curly braces behave the same as
editorconfig-core-c.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>